Service/requirements.txt
django==4.0.4
mysqlclient==2.1.0
pymysql==1.0.2
python-dotenv==0.20.0
djangorestframework==3.13.1
django-cors-headers==3.11.0